Rewrite main.rs for 10_macros_calling_macros
